feat(account-lib): js doc fixes
