+++ modules/account-lib/src/coin/eos/UnlinkAuthActionBuilder.ts
@@ -16,6 +16,13 @@ export class UnlinkAuthActionBuilder extends EosActionBuilder {
     this.action.name = this.actionName();
   }
 
+  /**
+   * Sets the permission's owner to be linked and the payer of the RAM needed to store
+   *
+   * @returns {this} this action builder.
+   *
+   * @param {string} account valid eos name
+   */
   account(account: string): this {
     if (Utils.isValidName(account)) {
       this._account = account;
@@ -23,6 +30,13 @@ export class UnlinkAuthActionBuilder extends EosActionBuilder {
     return this;
   }
 
+  /**
+   * Sets the owner of the action to be linked
+   *
+   * @returns {this} this action builder.
+   *
+   * @param {string} code valid eos name
+   */
   code(code: string): this {
     if (Utils.isValidName(code)) {
       this._code = code;
@@ -30,6 +44,13 @@ export class UnlinkAuthActionBuilder extends EosActionBuilder {
     return this;
   }
 
+  /**
+   * Sets the the action to be linked
+   *
+   * @returns {this} this action builder.
+   *
+   * @param {string} type valid eos name
+   */
   type(type: string): this {
     if (Utils.isValidName(type)) {
       this._type = type;
@@ -46,6 +67,7 @@ export class UnlinkAuthActionBuilder extends EosActionBuilder {
     return 'unlinkauth';
   }
 
+  /** @inheritdoc */
   build(builder: EosTxBuilder): EosJs.Serialize.Action {
     const data = this.action.data;
     if (typeof data === 'string') {
@@ -64,6 +86,13 @@ export class UnlinkAuthActionBuilder extends EosActionBuilder {
     }
   }
 
+  /**
+   * Validates whether the required fields are present
+   *
+   * @param {string} account name of account
+   * @param {string} code owner of the action to be linked
+   * @param {string} type the action to be linked
+   */
   private validateMandatoryFields(account: string, code: string, type: string) {
     const validationResult = UnlinkAuthActionSchema.validate({
       account,
