+++ modules/account-lib/src/coin/eos/eosTransactionBuilder.ts
@@ -2,13 +2,13 @@ import { BaseCoin as CoinConfig } from '@bitgo/statics';
 // import { TransactionType } from '../baseCoin';
 import { TransactionBuilder } from './transactionBuilder';
 import { Transaction } from './transaction';
-import { TransferActionBuilder } from './eosActionBuilder';
 import { StakeActionBuilder } from './StakeActionBuilder';
 import { UnstakeActionBuilder } from './UnstakeActionBuilder';
 import { UpdateAuthActionBuilder } from './UpdateAuthActionBuilder';
 import { DeleteAuthActionBuilder } from './DeleteAuthActionBuilder';
 import { LinkAuthActionBuilder } from './LinkAuthActionBuilder';
 import { UnlinkAuthActionBuilder } from './UnlinkAuthActionBuilder';
+import { TransferActionBuilder } from './TransferActionBuilder';
 
 export class EosTransactionBuilder extends TransactionBuilder {
   constructor(_coinConfig: Readonly<CoinConfig>) {
